处理未设置模式时，模式选择错误的问题

+++ frameworks/native/camera/src/input/camera_manager.cpp
@@ -234,7 +234,7 @@ sptr<CaptureSession> CameraManager::CreateCaptureSession(SceneMode mode)
         MEDIA_ERR_LOG("Failed to get capture session object from hcamera service!, %{public}d", retCode);
         return nullptr;
     }
-    captureSession->SetMode(static_cast<int32_t>(mode));
+    captureSession->SetMode(mode);
     return captureSession;
 }
